pyproject.toml

[project]
name = "pymavrest"
version = "1.0.0"
description = "Mavlink REST API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "build>=1.3.0",
    "cython>=3.2.1",
    "docker>=7.1.0",
    "fastapi==0.115.8",
    "folium>=0.19.5",
    "geojson>=3.2.0",
    "geopandas>=1.1.0",
    "geopy>=2.4.1",
    "httpx[http2]>=0.28.1",
    "ipykernel>=6.29.5",
    "loguru==0.7.3",
    "lxml>=5.3.2",
    "matplotlib>=3.10.3",
    "mavsdk>=3.0.1",
    "networkx>=3.5",
    "nuitka>=2.8.9",
    "numpy>=2.2.5",
    "opy-distbuilder>=0.9.1.1",
    "orjson>=3.10.18",
    "ping3>=4.0.8",
    "py-spy>=0.4.1",
    "pyarmor>=9.1.7",
    "pydantic-settings==2.8.0",
    "pyinstaller>=6.13.0",
    "pyjwt==2.10.1",
    "pymavlink>=2.4.43",
    "pyserial==3.5",
    "pyserial-asyncio>=0.6",
    "python-multipart==0.0.20",
    "pythonping==1.1.4",
    "pyulog>=1.2.2",
    "requests>=2.32.3",
    "rtree>=1.4.0",
    "scipy>=1.16.1",
    "setuptools>=78.1.0",
    "shapely>=2.1.0",
    "tenacity>=9.1.2",
    "typer==0.20",
    "uvicorn==0.34.0",
    "uvloop>=0.21.0",
    "websockets>=15.0.1",
]

[dependency-groups]
jupyter = [
    "ipykernel>=6.29.5",
]

[tool.setuptools]
packages = [
    "mavlink_rest",
    "mavlink_rest.repository",
]

